# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api documentation",
        "description": "Api documentation",
        "contact": {
            "email": "saguilarsanhuza@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.1"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/store": {
            "post": {
                "tags": [
                    "Stores"
                ],
                "summary": "Crear una nueva tienda",
                "operationId": "store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "image_store",
                                    "direction",
                                    "delivery_type_id",
                                    "actual_price",
                                    "old_price",
                                    "opening_hours",
                                    "closing_time",
                                    "logo_store"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Oxxo"
                                    },
                                    "image_store": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "/storage/images/image.jpg"
                                    },
                                    "direction": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "av los toros"
                                    },
                                    "delivery_type_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    },
                                    "actual_price": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "25000"
                                    },
                                    "old_price": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "25000"
                                    },
                                    "opening_hours": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "9:00"
                                    },
                                    "closing_time": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "21:00"
                                    },
                                    "logo_store": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "/storage/images/image.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Mensaje de exito al crear la tienda",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tienda creada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Mesaje de error en caso de que la tienda no es creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error al crear tienda"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/upload-file": {
            "post": {
                "tags": [
                    "Stores"
                ],
                "summary": "Cargar imagenes de la tienda",
                "operationId": "uploadFile",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "retornamos la url del archivo cargado"
                    }
                }
            }
        },
        "/store/{store}": {
            "get": {
                "tags": [
                    "Stores"
                ],
                "summary": "Recupera una tienda y sus productos",
                "description": "Recupera una tienda y sus productos",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "El ID de la tienda",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tienda recuperada con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "store": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tienda no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tienda no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Stores"
                ],
                "summary": "Actualizar tiendas",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "store",
                        "in": "path",
                        "description": "El ID de la tienda",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "image_store",
                                    "direction",
                                    "delivery_type_id",
                                    "actual_price",
                                    "old_price",
                                    "opening_hours",
                                    "closing_time",
                                    "logo_store"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Oxxo"
                                    },
                                    "image_store": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "/storage/images/image.jpg"
                                    },
                                    "direction": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "av los toros"
                                    },
                                    "delivery_type_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    },
                                    "actual_price": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "25000"
                                    },
                                    "old_price": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "25000"
                                    },
                                    "opening_hours": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "9:00"
                                    },
                                    "closing_time": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "21:00"
                                    },
                                    "logo_store": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "/storage/images/image.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tienda actualizada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tienda editada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al actualizar tienda",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error al editar tienda"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Stores"
                ],
                "summary": "Eliminar una tienda",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "store",
                        "in": "path",
                        "description": "El ID de la tienda",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Tienda eliminada con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tienda eliminada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al eliminar la tienda",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error al eliminar tienda"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}